Combine train_rllib.py and train_stable_baseline.py (#812)
* Make train.py single file

* flake8

* Fix remaining files

* Few bug fixes

* PR fix
