+++ examples/README.md
@@ -16,7 +16,9 @@ python simulate.py ring
 
 The examples are categorized into the following 3 sections:
 
-**non-RL examples** contains examples of transportation network with vehicles
+## non-RL examples 
+
+These are examples of transportation network with vehicles
 following human-dynamical models of driving behavior using the traffic 
 micro-simulator sumo and traffic macro-simulator Aimsun.
 
@@ -35,36 +37,43 @@ There are several *optional* arguments that can be added to the above command:
 ```
 where `--num_runs` indicates the number of simulations to run (default of `n` is 1), `--render` indicates whether to run the simulation during runtime (default is False), `--aimsun` indicates whether to run the simulation using the simulator Aimsun (the default simulator is SUMO), and `--gen_emission` indicates whether to generate an emission file from the simulation.
 
-**RL examples based on RLlib** provides similar networks as those presented in 
-the first point, but in the present of autonomous vehicle (AV) or traffic light agents 
+## RL examples based on RLlib 
+
+These examples are similar networks as those mentioned in *non-RL examples*, but in the 
+presence of autonomous vehicle (AV) or traffic light agents 
 being trained through RL algorithms provided by *RLlib*.
 
 To execute these examples, run
 
 ```shell
- python train_rllib.py EXP_CONFIG
+ python train.py EXP_CONFIG
+ (or python train.py EXP_CONFIG --rl_trainer RLlib)
 ```
 where `EXP_CONFIG` is the name of the experiment configuration file, as located in 
 `exp_configs/rl/singleagent` or  `exp_configs/rl/multiagent.`
 
 
-**RL examples based on stable_baselines** provides similar networks as those 
-presented in the first point, but in the present of autonomous vehicle (AV) or traffic 
+## RL examples based on stable_baselines 
+
+These examples provide similar networks as those 
+mentioned in *non-RL examples*, but in the presence of autonomous vehicle (AV) or traffic 
 light agents being trained through RL algorithms provided by OpenAI *stable 
 baselines*.
 
 To execute these examples, run
 
 ```shell
- python train_stable_baselines.py EXP_CONFIG
+ python train.py EXP_CONFIG --rl_trainer Stable-Baselines
 ```
 where `EXP_CONFIG` is the name of the experiment configuration file, as located in 
 `exp_configs/rl/singleagent.`
 
+Note that, currently, multiagent experiments are only supported through RLlib.
+
 There are several *optional* arguments that can be added to the above command:
 
 ```shell
- python train_stable_baselines.py EXP_CONFIG --num_cpus n1 --num_steps n2 --rollout_size r
+ python train.py EXP_CONFIG --rl_trainer Stable-Baselines --num_cpus n1 --num_steps n2 --rollout_size r
 ```
 where `--num_cpus` indicates the number of CPUs to use (default of `n1` is 1), `--num_steps` indicates the total steps to perform the learning (default of `n2` is 5000), and `--rollout_size` indicates the number of steps in a training batch (default of `r` is 1000)
